requirements.txt
python-dotenv
pyTelegramBotAPI
openai==0.28
